+++ wear/build.gradle
@@ -6,7 +6,7 @@ android {
     buildToolsVersion project.ANDROID_BUILD_TOOLS_VERSION
 
     defaultConfig {
-        applicationId "com.twotoasters.androidwearhelpers"
+        applicationId "com.ticketmaster.androidwearhelpers"
         minSdkVersion Integer.parseInt(project.MIN_SDK_VERSION_WEAR)
         targetSdkVersion Integer.parseInt(project.TARGET_SDK_VERSION)
         versionCode Integer.parseInt(project.VERSION_CODE_WEAR)
