+++ wear/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.twotoasters.androidwearhelpers" >
+    package="com.ticketmaster.androidwearhelpers" >
 
     <uses-feature android:name="android.hardware.type.watch" />
 
@@ -10,7 +10,7 @@
         android:label="@string/app_name"
         android:theme="@android:style/Theme.DeviceDefault" >
         <activity
-            android:name=".MyActivity"
+            android:name="com.ticketmaster.androidwearhelpers.MyActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
