--- wear/src/main/java/com/twotoasters/androidwearhelpers/MyActivity.java
@@ -1,24 +1,0 @@
-package com.twotoasters.androidwearhelpers;
-
-import android.app.Activity;
-import android.os.Bundle;
-import android.support.wearable.view.WatchViewStub;
-import android.widget.TextView;
-
-public class MyActivity extends Activity {
-
-    private TextView mTextView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_my);
-        final WatchViewStub stub = (WatchViewStub) findViewById(R.id.watch_view_stub);
-        stub.setOnLayoutInflatedListener(new WatchViewStub.OnLayoutInflatedListener() {
-            @Override
-            public void onLayoutInflated(WatchViewStub stub) {
-                mTextView = (TextView) stub.findViewById(R.id.text);
-            }
-        });
-    }
-}
